+++ src/application-navigation/ApplicationNavigation.jsx
@@ -9,6 +9,7 @@ import {
 
 import ApplicationErrorBoundary from '../application-error-boundary';
 import ApplicationLoadingOverlay, { ApplicationLoadingOverlayProvider } from '../application-loading-overlay';
+import { ApplicationStatusOverlayProvider } from '../application-status-overlay';
 import { NavigationPromptCheckpoint, navigationPromptResolutionOptionsShape, getUnsavedChangesPromptOptions } from '../navigation-prompt';
 import { ApplicationIntlContext } from '../application-intl';
 
@@ -183,15 +184,17 @@ const ApplicationNavigation = ({
       onDrawerMenuStateChange={onDrawerMenuStateChange}
     >
       <ApplicationLoadingOverlayProvider>
-        <NavigationPromptCheckpoint
-          ref={navigationPromptCheckpointRef}
-        >
-          <ApplicationErrorBoundary>
-            <Suspense fallback={<ApplicationLoadingOverlay isOpen />}>
-              {children}
-            </Suspense>
-          </ApplicationErrorBoundary>
-        </NavigationPromptCheckpoint>
+        <ApplicationStatusOverlayProvider>
+          <NavigationPromptCheckpoint
+            ref={navigationPromptCheckpointRef}
+          >
+            <ApplicationErrorBoundary>
+              <Suspense fallback={<ApplicationLoadingOverlay isOpen />}>
+                {children}
+              </Suspense>
+            </ApplicationErrorBoundary>
+          </NavigationPromptCheckpoint>
+        </ApplicationStatusOverlayProvider>
       </ApplicationLoadingOverlayProvider>
     </TerraApplicationNavigation>
   );